Updated Validations in the player model
After learning about validations in RoR I updated validations to
check for age and score entries that are too long.

To Do:  To validate numerically on the backend will require a schema
change so that data coming from the form as numbers and not strings.
Change needs to be made in HTML as well.